* added UA lemmatizer * choice lemmatize library in settings
