+++ parlacards/lemmatizers/lemmatizer_ua.py
@@ -1,0 +1,10 @@
+from uk_stemmer import UkStemmer
+
+
+# initialize the lemmatizer class only once
+stemmer = UkStemmer()
+def lemmatize(token):
+    return stemmer.stem_word(token)
+
+def lemmatize_many(tokens):
+    return [stemmer.stem_word(x) for x in tokens]
