+++ parladata/admin/admin.py
@@ -83,7 +83,7 @@ class SpeechForm(forms.ModelForm):
 
 class SpeechAdmin(admin.ModelAdmin):
     form = SpeechForm
-    fields = ['content', 'motions', 'speaker', 'order', 'tags', 'session']
+    fields = ['content', 'motions', 'speaker', 'order', 'tags', 'session', 'lemmatized_content']
     list_filter = ('session', 'tags')
     search_fields = ['speaker__name', 'content']
     autocomplete_fields = ['motions', 'speaker']
